move DetailsAndContents Page to a seperate file
