+++ magda-web-client/src/Components/Dataset/Add/DatasetAddMetadataPage.scss
@@ -61,99 +61,6 @@
         }
     }
 
-    .dataset-details-and-contents-page {
-        h2 {
-            font-size: 38px;
-            font-weight: normal;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 1.61;
-            letter-spacing: normal;
-            color: #484848;
-            margin-top: 60px;
-            margin-bottom: 80px;
-        }
-        h3 {
-            font-size: 29px;
-            font-weight: normal;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 1.03;
-            letter-spacing: normal;
-            color: #484848;
-            padding-bottom: 15px;
-        }
-
-        h4 {
-            font-size: 16px;
-            font-weight: normal;
-            font-style: normal;
-            font-stretch: normal;
-            line-height: 1.25;
-            letter-spacing: normal;
-            color: #484848;
-        }
-
-        .question-title {
-            margin-bottom: 58px;
-        }
-
-        .question-language {
-            margin-bottom: 72px;
-        }
-
-        .question-keyword {
-            margin-bottom: 50px;
-        }
-        .question-theme {
-            margin-bottom: 49px;
-        }
-
-        .question-description {
-            margin-bottom: 100px;
-            .tooltip-root {
-                max-width: 728px;
-            }
-        }
-
-        .question-issue-date {
-            h4 {
-                font-size: 16px;
-                font-weight: normal;
-                font-style: normal;
-                font-stretch: normal;
-                line-height: 1.25;
-                letter-spacing: normal;
-                color: #484848;
-                position: relative;
-            }
-        }
-
-        .question-recent-modify-date {
-            h4 {
-                font-size: 16px;
-                font-weight: normal;
-                font-style: normal;
-                font-stretch: normal;
-                line-height: 1.25;
-                letter-spacing: normal;
-                color: #484848;
-            }
-        }
-
-        .date-row {
-            margin-bottom: 68px;
-        }
-
-        .question-update-frequency {
-            margin-bottom: 68px;
-        }
-
-        .question-time-period {
-            margin-bottom: 50px;
-        }
-    }
-
     .dataset-submit-page {
         h2 {
             font-size: 38px;
